requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0.1
biopython>=1.81
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.65.0
pytest>=7.0.0
psutil>=5.9.0
